requirements.txt
Flask
Pillow
torch
torchvision
timm